+++ src/main/kotlin/instance/Instance.kt
@@ -26,6 +26,9 @@ data class Instance(
     val uses4GBPatch: Boolean,
     val isBaseInstance: Boolean = false
 ) : PersistentValue<SavedInstance> {
+    val manifest: Path get() = directory / "manifest.acf"
+    val manifestBackup: Path get() = directory / "manifest.backup.acf"
+
     suspend fun start(editor: Boolean = false, awaitSteamDeath: suspend () -> Boolean = { true }) {
         val flags = buildList {
             addAll(options.toLaunchFlags())
@@ -58,8 +61,6 @@ data class Instance(
             val VALUES = values().toList()
         }
     }
-
-
 
     override fun toSavedData(): SavedInstance {
         return SavedInstance(id, name, directory, patchVersion, options, uses4GBPatch, isBaseInstance)